+++ Module.php
@@ -6,6 +6,11 @@
  * @license https://www.humhub.com/licences
  */
 
+/**
+ *  Copyright (c) Ascensio System SIA 2022. All rights reserved.
+ *  http://www.onlyoffice.com
+ */
+
 namespace humhub\modules\onlyoffice;
 
 use Yii;
@@ -34,9 +39,9 @@ class Module extends \humhub\components\Module
     /**
      * Only document types
      */
-    const DOCUMENT_TYPE_TEXT = 'text';
-    const DOCUMENT_TYPE_PRESENTATION = 'presentation';
-    const DOCUMENT_TYPE_SPREADSHEET = 'spreadsheet';
+    const DOCUMENT_TYPE_TEXT = 'word';
+    const DOCUMENT_TYPE_PRESENTATION = 'slide';
+    const DOCUMENT_TYPE_SPREADSHEET = 'cell';
 
     /**
      * @var string[] allowed spreadsheet extensions 
@@ -101,6 +106,11 @@ class Module extends \humhub\components\Module
         return $this->settings->get('storageUrl');
     }
 
+    public function getVerifyPeerOff()
+    {
+        return $this->settings->get('verifyPeerOff');
+    }
+
     /**
      * 
      * @return type
@@ -112,7 +122,7 @@ class Module extends \humhub\components\Module
 
     public function getDocumentType($file)
     {
-        $fileExtension = FileHelper::getExtension($file);
+        $fileExtension = strtolower(FileHelper::getExtension($file));
 
         if (in_array($fileExtension, $this->spreadsheetExtensions)) {
             return self::DOCUMENT_TYPE_SPREADSHEET;
@@ -127,13 +137,13 @@ class Module extends \humhub\components\Module
 
     public function canEdit($file)
     {
-        $fileExtension = FileHelper::getExtension($file);
+        $fileExtension = strtolower(FileHelper::getExtension($file));
         return in_array($fileExtension, $this->editableExtensions);
     }
 
     public function canConvert($file)
     {
-        $fileExtension = FileHelper::getExtension($file);
+        $fileExtension = strtolower(FileHelper::getExtension($file));
         return in_array($fileExtension, $this->convertableExtensions);
     }
 
@@ -173,23 +183,19 @@ class Module extends \humhub\components\Module
         $url = $this->getInternalServerUrl() . '/coauthoring/CommandService.ashx';
 
         try {
-            $http = new \Zend\Http\Client($url, [
-                'adapter' => '\Zend\Http\Client\Adapter\Curl',
-                'curloptions' => CURLHelper::getOptions(),
-                'timeout' => 10
-            ]);
-            $http->setMethod('POST');
-            $headers = $http->getRequest()->getHeaders();
-
+            $headers = [];
+            $headers['Accept'] = 'application/json';
             if ($this->isJwtEnabled()) {
                 $data['token'] = JWT::encode($data, $this->getJwtSecret());
-                $headers->addHeaderLine('Authorization', 'Bearer ' . JWT::encode(['payload' => $data], $this->getJwtSecret()));
+                $headers['Authorization'] = 'Bearer ' . JWT::encode(['payload' => $data], $this->getJwtSecret());
             }
 
-            $http->setRawBody(Json::encode($data));
-            $headers->addHeaderLine('Accept', 'application/json');
+            $options = array(
+                'headers' => $headers,
+                'body' => $data
+            );
 
-            $response = $http->send();
+            $response = $this->request($url, 'POST', $options);
             $json = $response->getBody();
         } catch (\Exception $ex) {
             Yii::error('Could not get document server response! ' . $ex->getMessage());
@@ -209,34 +215,38 @@ class Module extends \humhub\components\Module
         $url = $this->getInternalServerUrl() . '/ConvertService.ashx';
         $key = $this->generateDocumentKey($file);
 
-        $ext = FileHelper::getExtension($file);
+        $user = Yii::$app->user->getIdentity();
+        $userGuid = null;
+        if (isset($user->guid)) {
+            $userGuid = $user->guid;
+        }
+
+        $docHash = $this->generateHash($key, $userGuid);
+
+        $ext = strtolower(FileHelper::getExtension($file));
         $data = [
             'async' => true,
             'embeddedfonts' => true,
             'filetype' => $ext,
             'outputtype' => $this->convertsTo[$ext],
             'key' => $key . $ts,
-            'url' => Url::to(['/onlyoffice/backend/download', 'key' => $key], true),
+            'url' => Url::to(['/onlyoffice/backend/download', 'doc' => $docHash], true),
         ];
 
         try {
-            $http = new \Zend\Http\Client($url, [
-                'adapter' => '\Zend\Http\Client\Adapter\Curl',
-                'curloptions' => CURLHelper::getOptions(),
-                'timeout' => 10
-            ]);
-            $http->setMethod('POST');
-            $headers = $http->getRequest()->getHeaders();
-
+            $headers = [];
+            $headers['Accept'] = 'application/json';
             if ($this->isJwtEnabled()) {
                 $data['token'] = JWT::encode($data, $this->getJwtSecret());
-                $headers->addHeaderLine('Authorization', 'Bearer ' . JWT::encode(['payload' => $data], $this->getJwtSecret()));
+                $headers['Authorization'] = 'Bearer ' . JWT::encode(['payload' => $data], $this->getJwtSecret());
             }
 
-            $http->setRawBody(Json::encode($data));
-            $headers->addHeaderLine('Accept', 'application/json');
+            $options = array(
+                'headers' => $headers,
+                'body' => $data
+            );
 
-            $response = $http->send();
+            $response = $this->request($url, 'POST', $options);
             $json = $response->getBody();
         } catch (\Exception $ex) {
             $error = 'Could not get document server response! ' . $ex->getMessage();
@@ -267,6 +277,37 @@ class Module extends \humhub\components\Module
         return [];
     }
 
+    /**
+     * @inheritdoc
+     */
+    public function generateHash($key, $userGuid)
+    {
+        $data = [
+            'key' => $key
+        ];
+
+        if (!empty($userGuid)) {
+            $data['userGuid'] = $userGuid;
+        }
+
+        return JWT::encode($data, Yii::$app->settings->get('secret'));
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function readHash($hash)
+    {
+        try {
+            $data = JWT::decode($hash, Yii::$app->settings->get('secret'), array('HS256'));
+        } catch (\Exception $ex) {
+            $error = 'Invalid hash ' . $ex->getMessage();
+            Yii::error($error);
+            return [null, $error];
+        }
+
+        return [$data, null];
+    }
     /**
      * @inheritdoc
      */