+++ messages/fr/base.php
@@ -53,5 +53,6 @@ return [
     'Allows the user to use ONLYOFFICE.' => 'Permet à l\'utilisateur d\'utiliser ONLYOFFICE.',
     'Form template' => 'Modèle de formulaire',
     'Fill in form in ONLYOFFICE' => 'Remplir le formulaire dans ONLYOFFICE',
-    'Please update ONLYOFFICE Docs to version 7.0 to work on fillable forms online' => 'Veuillez mettre à jour ONLYOFFICE Docs vers la version 7.0 pour travailler sur les formulaires à remplir en ligne'
+    'Please update ONLYOFFICE Docs to version 7.0 to work on fillable forms online' => 'Veuillez mettre à jour ONLYOFFICE Docs vers la version 7.0 pour travailler sur les formulaires à remplir en ligne',
+    'Disable certificate verification (insecure)' => 'Désactiver la vérification du certificat (non sur)',
 ];