+++ messages/pl/base.php
@@ -53,5 +53,6 @@ return [
     'Allows the user to use ONLYOFFICE.' => 'Zezwala użytkownikowi na korzystanie z ONLYOFFICE.',
     'Form template' => 'Szablon formularza',
     'Fill in form in ONLYOFFICE' => 'Wypełnić formularz w ONLYOFFICE',
-    'Please update ONLYOFFICE Docs to version 7.0 to work on fillable forms online' => 'Zaktualizuj ONLYOFFICE Docs do wersji 7.0, aby działały w formularzach do wypełniania online'
+    'Please update ONLYOFFICE Docs to version 7.0 to work on fillable forms online' => 'Zaktualizuj ONLYOFFICE Docs do wersji 7.0, aby działały w formularzach do wypełniania online',
+    'Disable certificate verification (insecure)' => 'Wyłącz weryfikację certyfikatu (niebezpieczne)',
 ];