+++ resources/humhub.onlyoffice.js
@@ -1,3 +1,8 @@
+/**
+ *  Copyright (c) Ascensio System SIA 2022. All rights reserved.
+ *  http://www.onlyoffice.com
+ */
+
 humhub.module('onlyoffice', function (module, require, $) {
 
     var client = require('client');
@@ -23,18 +28,17 @@ humhub.module('onlyoffice', function (module, require, $) {
     };
 
     Editor.prototype.init = function () {
-        
+        this.modal = modal.get('#onlyoffice-modal');
+
         if (this.options.moduleConfigured != 1) {
             module.log.error('No OnlyOffice server configured! - Check onlyoffice module configuration!', true);
             return
         }
-        
+
         this.initEditor();
-        
-        this.modal = modal.get('#onlyoffice-modal');
-        
+
         var that = this;
-        this.modal.$.on('hidden.bs.modal', function(evt) {
+        this.modal.$.on('hidden.bs.modal', function (evt) {
             that.modal.clear();
         });
 
@@ -50,24 +54,26 @@ humhub.module('onlyoffice', function (module, require, $) {
 
     Editor.prototype.close = function (evt) {
 
-        if (this.options.editMode == 'edit') {
-            refreshFileInfo(this, evt);
-        } else {
-            if (this.docEditor) {
-                this.docEditor.destroyEditor();
+        if (this.docEditor
+            && this.options.editMode == 'edit') {
+            if (this.docEditor.requestClose) {
+                setTimeout(() => {
+                    onRequestCloseObj = { that: this, evt: evt };
+                    this.docEditor.requestClose();
+                    evt.finish();
+                }, 0);
+            } else {
+                refreshFileInfo(this, evt);
             }
-            this.modal.clear();
-            this.modal.close();
-            evt.finish();
+        } else {
+            closeModal(this, evt);
         }
-        
-        
     }
 
     Editor.prototype.initEditor = function () {
-        if(!window.DocsAPI) {
+        if (!window.DocsAPI) {
             ooJSLoadRetries++;
-            if(ooJSLoadRetries < 100) {
+            if (ooJSLoadRetries < 100) {
                 setTimeout($.proxy(this.initEditor, this), 100);
                 return;
             } else {
@@ -75,13 +81,19 @@ humhub.module('onlyoffice', function (module, require, $) {
                 return;
             }
         }
-        
+
         api = this.options.api;
 
         var config = this.options.config;
 
+        var docsVersion = DocsAPI.DocEditor.version().split(".");
+        if (docsVersion[0] < 6
+            || docsVersion[0] == 6 && docsVersion[1] == 0) {
+            module.log.error('Not supported version', true);
+            return;
+        }
         if ((config.document.fileType === 'docxf' || config.document.fileType === 'oform')
-            && DocsAPI.DocEditor.version().split('.')[0] < 7) {
+            && docsVersion[0] < 7) {
             module.log.error('Please update ONLYOFFICE Docs to version 7.0 to work on fillable forms online', true);
             return;
         }
@@ -89,6 +101,7 @@ humhub.module('onlyoffice', function (module, require, $) {
         config.width = "100%";
         config.height = "100%";
         config.events = {
+            'onRequestClose': onRequestClose,
             //'onReady': onReady,
             //'onDocumentStateChange': onDocumentStateChange,
             //'onRequestEditRights': onRequestEditRights,
@@ -120,10 +133,10 @@ humhub.module('onlyoffice', function (module, require, $) {
 
         function _callAjax() {
             jQuery.ajax({
-                type : "POST",
-                url : that.options.convertPost,
+                type: "POST",
+                url: that.options.convertPost,
                 cache: false,
-                success: function(data) {
+                success: function (data) {
                     if (data.error) {
                         _onError(data.error);
                         return;
@@ -145,7 +158,7 @@ humhub.module('onlyoffice', function (module, require, $) {
         }
 
         loader.set(that.$.find('.modal-footer'));
-		_callAjax();
+        _callAjax();
     };
 
     Convert.prototype.getDefaultOptions = function () {
@@ -169,30 +182,33 @@ humhub.module('onlyoffice', function (module, require, $) {
         });
     }
 
+    var onRequestCloseObj = null;
+    function onRequestClose() {
+        refreshFileInfo(onRequestCloseObj.that, onRequestCloseObj.evt);
+    };
+
     function refreshFileInfo(that, evt) {
-        client.post({url: that.options.fileInfoUrl}).then(function (response) {
+        client.post({ url: that.options.fileInfoUrl }).then(function (response) {
             event.trigger('humhub:file:modified', [response.file]);
-            if (that.docEditor) {
-                that.docEditor.destroyEditor();
-            }
-            if (that.modal) {
-                that.modal.clear();
-                that.modal.close();
-            }
-            evt.finish();
         }).catch(function (e) {
-            if (that.docEditor) {
-                that.docEditor.destroyEditor();
-            }
             module.log.error(e);
-            if (that.modal) {
-                that.modal.clear();
-                that.modal.close();
-            }
-            evt.finish();
+        }).finally(function () {
+            closeModal(that, evt);
         });
     }
 
+    function closeModal(that, evt) {
+        if (that.docEditor) {
+            that.docEditor.destroyEditor();
+        }
+        if (that.modal) {
+            that.modal.clear();
+            that.modal.close();
+        }
+        if (evt && evt.finish) {
+            evt.finish();
+        }
+    }
 
     var Share = function (node, options) {
         Widget.call(this, node, options);
@@ -225,7 +241,7 @@ humhub.module('onlyoffice', function (module, require, $) {
                     url: that.options.shareGetLink,
                     cache: false,
                     type: 'POST',
-                    data: {'shareMode': 'view'},
+                    data: { 'shareMode': 'view' },
                     dataType: 'json',
                     success: function (json) {
                         $('.viewLinkInput').show();
@@ -239,7 +255,7 @@ humhub.module('onlyoffice', function (module, require, $) {
                     url: that.options.shareRemoveLink,
                     cache: false,
                     type: 'POST',
-                    data: {'shareMode': 'view'},
+                    data: { 'shareMode': 'view' },
                     dataType: 'json',
                     success: function (jsoin) {
                         $('.viewLinkInput').hide();
@@ -256,7 +272,7 @@ humhub.module('onlyoffice', function (module, require, $) {
                     url: that.options.shareGetLink,
                     cache: false,
                     type: 'POST',
-                    data: {'shareMode': 'edit'},
+                    data: { 'shareMode': 'edit' },
                     dataType: 'json',
                     success: function (json) {
                         $('.editLinkInput').show();
@@ -270,7 +286,7 @@ humhub.module('onlyoffice', function (module, require, $) {
                     url: that.options.shareRemoveLink,
                     cache: false,
                     type: 'POST',
-                    data: {'shareMode': 'edit'},
+                    data: { 'shareMode': 'edit' },
                     dataType: 'json',
                     success: function (jsoin) {
                         $('.editLinkInput').hide();
@@ -293,9 +309,13 @@ humhub.module('onlyoffice', function (module, require, $) {
 
 
 
-    var init = function (pjax) {};
+    var init = function (pjax) { };
 
     var createSubmit = function (evt) {
+        if ($('#cfiles-folderView').length > 0) {
+            $('input#createdocument-fid').val($('#cfiles-folderView').attr('data-fid'));
+        }
+
         client.submit(evt).then(function (response) {
             event.trigger('humhub:file:created', [response.file]);