+++ views/create/document.php
@@ -1,7 +1,10 @@
 <?php
 
-use humhub\modules\onlyoffice\Module;
-use yii\helpers\Url;
+/**
+ *  Copyright (c) Ascensio System SIA 2022. All rights reserved.
+ *  http://www.onlyoffice.com
+ */
+
 use humhub\widgets\ActiveForm;
 use humhub\libs\Html;
 
@@ -17,6 +20,8 @@ $modal = \humhub\widgets\ModalDialog::begin([
 <div class="modal-body">
     <?= $form->field($model, 'fileName', ['template' => '{label}<div class="input-group">{input}<div class="input-group-addon">' . $model->extension . '</div></div>{hint}{error}']); ?>
     <?= $form->field($model, 'openFlag')->checkbox(); ?>
+
+    <?= $form->field($model, 'fid')->hiddenInput()->label(false); ?>
 </div>
 
 <div class="modal-footer">