+++ widgets/ShareWidget.php
@@ -6,10 +6,13 @@
  * @license https://www.humhub.com/licences
  */
 
+/**
+ *  Copyright (c) Ascensio System SIA 2022. All rights reserved.
+ *  http://www.onlyoffice.com
+ */
+
 namespace humhub\modules\onlyoffice\widgets;
 
-use Yii;
-use yii\web\HttpException;
 use yii\helpers\Url;
 use humhub\modules\file\models\File;
 use humhub\modules\onlyoffice\models\Share;